pyproject.toml
[project]
name = "nomenclature-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Kseniia Kornilova",email = "ksyu.kornilova2809@gmail.com"}
]
readme = "README.md"
requires-python = "3.11.*"
dependencies = [
    "django (==4.2)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-mptt (>=0.18.0,<0.19.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
isort = "^7.0.0"
flake8 = "^7.3.0"
pre-commit = "^4.3.0"
flake8-bugbear = "^25.10.21"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"
faker = "^37.12.0"
pytest-factoryboy = "^2.8.1"
pytest = "^8.4.2"
factory-boy = "^3.3.3"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["backend", "shop"]